XMLHttpRequest/open-url-fragment.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() resolving URLs - fragment identifier</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-open()-method" data-tested-assertations="following::ol[1]/li[7]" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "folder.txt#foobar", false)
        client.send(null)
        assert_equals(client.responseText, "top\n")
      })
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "resources/requri.php#foobar", false)
        client.send(null)
        assert_regexp_match(client.responseText, /XMLHttpRequest\/resources\/requri\.php$/)
      }, 'make sure fragment is removed from URL before request')
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "resources/requri.php?help=#foobar", false)
        client.send(null)
        assert_regexp_match(client.responseText, /XMLHttpRequest\/resources\/requri\.php\?help=$/)
      }, 'make sure fragment is removed from URL before request (with query string)')
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "resources/requri.php?" +encodeURIComponent("#foobar"), false)
        client.send(null)
        assert_regexp_match(client.responseText, /XMLHttpRequest\/resources\/requri\.php\?%23foobar$/)
      }, 'make sure escaped # is not removed')
    </script>
  </body>
</html>
